Adjust exports in p2panda-sync
